requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.2.1
Django==5.2.3
django-environ==0.12.0
django-model-utils==5.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
executing==2.2.0
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.9.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.0
wcwidth==0.2.13
pytest==8.4.1
pytest-django==4.11.1
